// src/pages/types/layoutTable.types.ts
export interface IOrderBy {
  key: string
  order: 'asc' | 'desc'
}